refactor: all confusing operation/instruction terminology replaced with "task"
