+++ README.rdoc
@@ -778,9 +778,9 @@ Adds class and instance methods for basic collision detection.
   
 ==== Trait "asynchronous"
 Allows your code to specify a GameObject's behavior asynchronously, including
-tweening, movement and even method calls. Instructions are added to a queue to
-be processed in order; the operation at the front of the queue is updated each
-tick and removed when it has finished.
+tweening, movement and even method calls. Tasks are added to a queue to be
+processed in order; the task at the front of the queue is updated each tick
+and removed when it has finished.
 
   # Simple one-trick example
   # This will cause an object to move from its current location to 64,64.
@@ -796,7 +796,7 @@ tick and removed when it has finished.
     q.call :destroy
   end
 
-Currently available operations are wait(timeout, &condition), tween(timeout,
+Currently available tasks are wait(timeout, &condition), tween(timeout,
 properties), call(method, *arguments) and exec { ... }.
 
 For a more complete example of how to use this trait, see
