+++ lib/chingu.rb
@@ -35,7 +35,7 @@ require File.join(CHINGU_ROOT,"chingu","require_all") # Thanks to http://github.
 require_all "#{CHINGU_ROOT}/chingu/helpers"
 require_all "#{CHINGU_ROOT}/chingu/traits"
 require_all "#{CHINGU_ROOT}/chingu/async"
-require_all "#{CHINGU_ROOT}/chingu/async_operations"
+require_all "#{CHINGU_ROOT}/chingu/async_tasks"
 require_all "#{CHINGU_ROOT}/chingu"
 
 module Chingu
